refactor: guards 폴더에 index 추가
